% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mybin}
\alias{mybin}
\title{binomial sampling function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterations}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
returns a graph of the simulation with proportions and a table of proportions.
}
\description{
This uses a sample to make a binomial simulation.
}
\examples{
mybin(iter=1000,n=30,p=0.7)
}
