% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mychisim}
\alias{mychisim}
\title{My chi-squared}
\usage{
mychisim(n1 = 10, sigma1 = 3, mean1 = 5, iter = 1000, ymax = 0.1)
}
\arguments{
\item{n1}{sample size}

\item{sigma1}{standard deviation}

\item{mean1}{mean}

\item{iter}{iterations}

\item{ymax}{max y value}
}
\value{
Chi-squared graph of the values with a theoretical line and simulated line
}
\description{
This function returns a chi-squared graph with a theoretical and simulated line.
}
\examples{
mychisim(30,4,5,10000,0.15)
}
