% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mycltu}
\alias{mycltu}
\title{central limit tendency uniform function}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{size of sample}

\item{iter}{iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
a histogram with a theoretical and simulated line to show density curve
}
\description{
This function returns a CLT graph with theoretical and simulated density curves.
}
\examples{
mycltu(10,100)
}
