% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mysquarefun}
\alias{mysquarefun}
\title{Squares the entries of a vector}
\usage{
mysquarefun(x)
}
\arguments{
\item{x}{a vector for squaring}
}
\value{
returns squared vector
}
\description{
This function returns the square of a vector.
}
\examples{
mysquarefun(someVector)
}
