% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pareto}
\alias{pareto}
\title{Generates a pareto graph}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{vector of data to use}

\item{mn}{title of barplot}

\item{...}{whatever else you want to add to the title}
}
\value{
a pareto graph
}
\description{
This function creates a pareto graph from the given vector x.
}
\examples{
pareto(x1,mn="graph")
}
